Remove old tests
