+++ tests/test.js
@@ -1,87 +1,687 @@
-/*
- * Copyright (C) 2013-2016 Trent Houliston <trent@houliston.me>, Jake Woods <jake.f.woods@gmail.com>
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
- * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the
- * Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE
- * WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
+const { test } = require('uvu');
+const assert = require('uvu/assert');
 
-var NUClearNet = require('../index.js').NUClearNet;
+const { NUClearNet } = require('..');
 
-var SegfaultHandler = require('segfault-handler');
-SegfaultHandler.registerHandler("crash.log");
-
-var name = 'nuclearnet_js_test'
-
-if (process.argv.length > 2) {
-    name = process.argv[2]
+function uniqueId() {
+  return String(Math.random() * 100000000).slice(0, 8);
 }
-var nu = new NUClearNet();
 
-nu.on('nuclear_join', function (peer) {
-    console.log('Join', peer);
+function asyncTest(testFn, { timeout = 1000 } = {}) {
+  return new Promise((resolve, reject) => {
+    let cleanUp;
+    let timeoutId;
 
-    console.log('Send', peer.name, 'Unreliable');
-    nu.send({
-        type: stringName,
-        payload: Buffer.from('Hello World from Javascript! (targeted UNRELIABLE)'),
-        target: peer.name,
-    });
-
-    console.log('Send', peer.name, 'Reliable');
-    nu.send({
-        type: stringName,
-        payload: Buffer.from('Hello World from Javascript! (targeted RELIABLE)'),
-        target: peer.name,
-        reliable: true,
-    });
-});
-
-nu.on('nuclear_leave', function (peer) {
-    console.log('Leave', peer);
-});
-
-var stringName = 'std::__1::basic_string<char, std::__1::char_traits<char>, std::__1::allocator<char> >';
-
-nu.on(stringName, function(packet) {
-    console.log(packet.reliable ? 'Reliable' : 'Unreliable', 'message from', packet.peer.name);
-    var string = packet.payload.toString();
-
-    if(string.length < 100) {
-        console.log(string);
+    if (timeout > 0) {
+      timeoutId = setTimeout(() => {
+        timeoutId = null;
+        cleanUp && cleanUp();
+        reject(new Error(`async test timed out after ${timeout}ms`));
+      }, timeout);
     }
-    else {
-        console.log(string[0]);
+
+    function done() {
+      if (timeoutId) {
+        clearTimeout(timeoutId);
+        timeoutId = null;
+      }
+
+      resolve();
     }
+
+    function fail(reason) {
+      if (timeoutId) {
+        clearTimeout(timeoutId);
+        timeoutId = null;
+      }
+
+      reject(new Error(reason));
+    }
+
+    cleanUp = testFn(done, fail);
+  });
+}
+
+test('NUClearNet instance throws if used after destroy()', () => {
+  const net = new NUClearNet();
+  net.destroy();
+
+  assert.throws(
+    () => {
+      net.on('some-event', () => {});
+    },
+    /This network instance has been destroyed/,
+    'NUClearNet.on() throws if called after instance is destroyed'
+  );
+
+  assert.throws(
+    () => {
+      net.addListener('some-event', () => {});
+    },
+    /This network instance has been destroyed/,
+    'NUClearNet.addListener() throws if called after instance is destroyed'
+  );
+
+  assert.throws(
+    () => {
+      net.hash('nuclearnet');
+    },
+    /This network instance has been destroyed/,
+    'NUClearNet.hash() throws if called after instance is destroyed'
+  );
+
+  assert.throws(
+    () => {
+      net.connect({});
+    },
+    /This network instance has been destroyed/,
+    'NUClearNet.connect() throws if called after instance is destroyed'
+  );
+
+  assert.throws(
+    () => {
+      net.disconnect();
+    },
+    /This network instance has been destroyed/,
+    'NUClearNet.disconnect() throws if called after instance is destroyed'
+  );
+
+  assert.throws(
+    () => {
+      net.send({});
+    },
+    /This network instance has been destroyed/,
+    'NUClearNet.send() throws if called after instance is destroyed'
+  );
 });
 
-nu.connect({
-    name: name,
+test('NUClearNet.hash()', () => {
+  const net = new NUClearNet();
+
+  const expected = Buffer.from('71226911ef5289e4', 'hex');
+  const hash = net.hash('nuclearnet');
+
+  assert.equal(hash, expected, 'Hash of "nuclearnet" matches known hash');
+
+  net.destroy();
 });
 
-process.on('SIGINT', () => {
-    console.log('Disconnecting');
-    nu.disconnect();
+test('NUClearNet.send() throws if used before connect()', () => {
+  const net = new NUClearNet();
+
+  assert.throws(() => {
+    net.send('nuclearnet');
+  }, /The network is not currently connected/);
+
+  net.destroy();
 });
 
-console.log('Sending All Unreliable');
-nu.send({
-    type: stringName,
-    payload: Buffer.from('Hello World from Javascript! (ALL UNRELIABLE)'),
+test('NUClearNet emits join events', async () => {
+  // Test set up:
+  //   - Create two network instances (A and B) and connect them
+  //   - Wait for A to join B (and vice versa) by listening for the `nuclear_join` event on both instances
+  //   - End successfully if A joined B, and B joined A
+  //   - Automatically end with failure if at least one didn't join the other before the timeout
+  await asyncTest(
+    (done) => {
+      const netA = new NUClearNet();
+      const netB = new NUClearNet();
+
+      function cleanUp() {
+        netA.destroy();
+        netB.destroy();
+      }
+
+      let aJoinedB = false;
+      let bJoinedA = false;
+
+      function checkComplete() {
+        if (aJoinedB && bJoinedA) {
+          cleanUp();
+          done();
+        }
+      }
+
+      netA.on('nuclear_join', (peer) => {
+        if (peer.name === netB.options.name) {
+          bJoinedA = true;
+          checkComplete();
+        }
+      });
+
+      netB.on('nuclear_join', (peer) => {
+        if (peer.name === netA.options.name) {
+          aJoinedB = true;
+          checkComplete();
+        }
+      });
+
+      netA.connect({
+        name: 'netA-' + uniqueId(),
+      });
+
+      netB.connect({
+        name: 'netB-' + uniqueId(),
+      });
+
+      return cleanUp;
+    },
+    { timeout: 1000 }
+  );
 });
 
-console.log('Sending All Reliable');
-nu.send({
-    type: stringName,
-    payload: Buffer.from('Hello World from Javascript! (ALL RELIABLE)'),
-    reliable: true,
+test('NUClearNet emits leave events', async () => {
+  // Test set up:
+  //   - Create two network instances (A and B) and connect them
+  //   - Wait for B to join A, then disconnect B to trigger the `nuclear_leave` event on A
+  //   - End successfully if B triggered the leave event after joining and disconnecting from A
+  //   - Automatically end with failure if the above didn't happen before the timeout
+  await asyncTest(
+    (done) => {
+      const netA = new NUClearNet();
+      const netB = new NUClearNet();
+
+      function cleanUp() {
+        netA.destroy();
+        netB.destroy();
+      }
+
+      netA.on('nuclear_leave', (peer) => {
+        if (peer.name === netB.options.name) {
+          cleanUp();
+          done();
+        }
+      });
+
+      netA.on('nuclear_join', (peer) => {
+        // Disconnect B after it joins, to trigger the leave event on A
+        if (peer.name === netB.options.name) {
+          netB.disconnect();
+        }
+      });
+
+      netA.connect({
+        name: 'netA-' + uniqueId(),
+      });
+
+      netB.connect({
+        name: 'netB-' + uniqueId(),
+      });
+
+      return cleanUp;
+    },
+    { timeout: 1000 }
+  );
 });
+
+test('NUClearNet can send and receive reliable targeted messages', async () => {
+  // Test set up:
+  //   - Create three network instances (A, B, C) and connect them
+  //   - Wait for B join A, then send B a specific payload with `reliable` set
+  //   - Do the above for C as well, with its own specific payload
+  //   - End successfully if both B and C got their respective payloads from A
+  //   - End with failure if B or C got a payload from A that was not their own (this verifies targeting works)
+  //   - Automatically end with failure if all of the above didn't happen before the timeout
+  await asyncTest(
+    (done, fail) => {
+      const netA = new NUClearNet();
+      const netB = new NUClearNet();
+      const netC = new NUClearNet();
+
+      function cleanUp() {
+        netA.destroy();
+        netB.destroy();
+        netC.destroy();
+      }
+
+      const payloadToB = Buffer.from('oh hai b!');
+      const payloadToC = Buffer.from('oh hai c!');
+
+      netA.on('nuclear_join', (peer) => {
+        // Send message to B after it joins
+        if (peer.name === netB.options.name) {
+          netA.send({
+            target: netB.options.name,
+            reliable: true,
+            type: 'message-from-a',
+            payload: payloadToB,
+          });
+        }
+
+        // Send message to C after it joins
+        if (peer.name === netC.options.name) {
+          netA.send({
+            target: netC.options.name,
+            reliable: true,
+            type: 'message-from-a',
+            payload: payloadToC,
+          });
+        }
+      });
+
+      let bGotMessage = false;
+      let cGotMessage = false;
+
+      function checkComplete() {
+        if (bGotMessage && cGotMessage) {
+          cleanUp();
+          done();
+        }
+      }
+
+      netB.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payloadToB) === 0) {
+            bGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netB received unexpected data from netA' +
+                '\n  expected: ' +
+                payloadToB.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf-8')
+            );
+          }
+        }
+      });
+
+      netC.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payloadToC) === 0) {
+            cGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netC received unexpected data from netA' +
+                '\n  expected: ' +
+                payloadToC.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf8')
+            );
+          }
+        }
+      });
+
+      netA.connect({
+        name: 'netA-' + uniqueId(),
+      });
+
+      netB.connect({
+        name: 'netB-' + uniqueId(),
+      });
+
+      netC.connect({
+        name: 'netC-' + uniqueId(),
+      });
+
+      return cleanUp;
+    },
+    { timeout: 1000 }
+  );
+});
+
+test('NUClearNet can send and receive unreliable targeted messages', async () => {
+  // Test set up:
+  //   - Create three network instances (A, B, C) and connect them
+  //   - Wait for B to join A, then start an interval to unreliably send B a specific payload
+  //     The interval ensures multiple messages will be sent, to compensate for any that are dropped due to the unreliable send.
+  //   - Do the above for C as well, with its own specific payload
+  //   - End successfully if both B and C got their respective payloads from A
+  //   - End with failure if B or C got a payload from A that was not their own (this verifies targeting works)
+  //   - Automatically end with failure if all of the above didn't happen before the timeout
+  await asyncTest(
+    (done, fail) => {
+      const netA = new NUClearNet();
+      const netB = new NUClearNet();
+      const netC = new NUClearNet();
+
+      let bSendInterval;
+      let cSendInterval;
+
+      function cleanUp() {
+        if (bSendInterval) {
+          clearInterval(bSendInterval);
+          bSendInterval = undefined;
+        }
+
+        if (cSendInterval) {
+          clearInterval(cSendInterval);
+          cSendInterval = undefined;
+        }
+
+        netA.destroy();
+        netB.destroy();
+        netC.destroy();
+      }
+
+      const payloadToB = Buffer.from('oh hai b!');
+      const payloadToC = Buffer.from('oh hai c!');
+
+      netA.on('nuclear_join', (peer) => {
+        // Start sending unreliable messages to B after it joins
+        if (peer.name === netB.options.name) {
+          bSendInterval && clearInterval(bSendInterval);
+          bSendInterval = setInterval(() => {
+            netA.send({
+              target: netB.options.name,
+              reliable: false,
+              type: 'message-from-a',
+              payload: payloadToB,
+            });
+          }, 50);
+        }
+
+        // Start sending unreliable messages to C after it joins
+        if (peer.name === netC.options.name) {
+          cSendInterval && clearInterval(cSendInterval);
+          cSendInterval = setInterval(() => {
+            netA.send({
+              target: netC.options.name,
+              reliable: false,
+              type: 'message-from-a',
+              payload: payloadToC,
+            });
+          }, 50);
+        }
+      });
+
+      let bGotMessage = false;
+      let cGotMessage = false;
+
+      function checkComplete() {
+        if (bGotMessage && cGotMessage) {
+          cleanUp();
+          done();
+        }
+      }
+
+      netB.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payloadToB) === 0) {
+            bGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netB received unexpected data from netA' +
+                '\n  expected: ' +
+                payloadToB.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf-8')
+            );
+          }
+        }
+      });
+
+      netC.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payloadToC) === 0) {
+            cGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netC received unexpected data from netA' +
+                '\n  expected: ' +
+                payloadToC.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf8')
+            );
+          }
+        }
+      });
+
+      netA.connect({
+        name: 'netA-' + uniqueId(),
+      });
+
+      netB.connect({
+        name: 'netB-' + uniqueId(),
+      });
+
+      netC.connect({
+        name: 'netC-' + uniqueId(),
+      });
+
+      return cleanUp;
+    },
+    { timeout: 1000 }
+  );
+});
+
+test('NUClearNet can send and receive reliable untargeted messages', async () => {
+  // Test set up:
+  //   - Create three network instances (A, B, C) and connect them
+  //   - Wait for both B and C to join A, then send the payload with `reliable` set, untargeted
+  //   - End successfully if both B and C got the payload from A
+  //   - End with failure if B or C got a payload from A that was not the expected payload
+  //   - Automatically end with failure if all of the above didn't happen before the timeout
+  await asyncTest(
+    (done, fail) => {
+      const netA = new NUClearNet();
+      const netB = new NUClearNet();
+      const netC = new NUClearNet();
+
+      function cleanUp() {
+        netA.destroy();
+        netB.destroy();
+        netC.destroy();
+      }
+
+      const payload = Buffer.from('oh hai guys!');
+
+      let bConnected = false;
+      let cConnected = false;
+
+      function checkConnected() {
+        if (bConnected && cConnected) {
+          netA.send({
+            reliable: true,
+            type: 'message-from-a',
+            payload,
+          });
+        }
+      }
+
+      netA.on('nuclear_join', (peer) => {
+        if (peer.name === netB.options.name) {
+          bConnected = true;
+          checkConnected();
+        }
+
+        if (peer.name === netC.options.name) {
+          cConnected = true;
+          checkConnected();
+        }
+      });
+
+      let bGotMessage = false;
+      let cGotMessage = false;
+
+      function checkComplete() {
+        if (bGotMessage && cGotMessage) {
+          cleanUp();
+          done();
+        }
+      }
+
+      netB.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payload) === 0) {
+            bGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netB received unexpected data from netA' +
+                '\n  expected: ' +
+                payload.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf-8')
+            );
+          }
+        }
+      });
+
+      netC.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payload) === 0) {
+            cGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netC received unexpected data from netA' +
+                '\n  expected: ' +
+                payload.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf8')
+            );
+          }
+        }
+      });
+
+      netA.connect({
+        name: 'netA-' + uniqueId(),
+      });
+
+      netB.connect({
+        name: 'netB-' + uniqueId(),
+      });
+
+      netC.connect({
+        name: 'netC-' + uniqueId(),
+      });
+
+      return cleanUp;
+    },
+    { timeout: 5000 }
+  );
+});
+
+test('NUClearNet can send and receive unreliable untargeted messages', async () => {
+  // Test set up:
+  //   - Create three network instances (A, B, C) and connect them
+  //   - Wait for both B and C to join A, then start an interval to unreliably send the same payload, without a target.
+  //     The interval ensures multiple messages will be sent, to compensate for any that are dropped due to the unreliable send.
+  //   - End successfully if both B and C got the payload from A
+  //   - End with failure if B or C got a payload from A that was not the expected payload
+  //   - Automatically end with failure if all of the above didn't happen before the timeout
+  await asyncTest(
+    (done, fail) => {
+      const netA = new NUClearNet();
+      const netB = new NUClearNet();
+      const netC = new NUClearNet();
+
+      let sendInterval;
+
+      function cleanUp() {
+        if (sendInterval) {
+          clearInterval(sendInterval);
+          sendInterval = undefined;
+        }
+
+        netA.destroy();
+        netB.destroy();
+        netC.destroy();
+      }
+
+      const payload = Buffer.from('oh hai guys!');
+
+      let bConnected = false;
+      let cConnected = false;
+
+      function checkConnected() {
+        // Start sending unreliable messages after both peers are connected
+        if (bConnected && cConnected) {
+          sendInterval && clearInterval(sendInterval);
+          sendInterval = setInterval(() => {
+            netA.send({
+              reliable: false,
+              type: 'message-from-a',
+              payload,
+            });
+          }, 50);
+        }
+      }
+
+      netA.on('nuclear_join', (peer) => {
+        if (peer.name === netB.options.name) {
+          bConnected = true;
+          checkConnected();
+        }
+
+        if (peer.name === netC.options.name) {
+          cConnected = true;
+          checkConnected();
+        }
+      });
+
+      let bGotMessage = false;
+      let cGotMessage = false;
+
+      function checkComplete() {
+        if (bGotMessage && cGotMessage) {
+          cleanUp();
+          done();
+        }
+      }
+
+      netB.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payload) === 0) {
+            bGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netB received unexpected data from netA' +
+                '\n  expected: ' +
+                payload.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf-8')
+            );
+          }
+        }
+      });
+
+      netC.on('message-from-a', (packet) => {
+        if (packet.peer.name === netA.options.name) {
+          if (packet.payload.compare(payload) === 0) {
+            cGotMessage = true;
+            checkComplete();
+          } else {
+            cleanUp();
+            fail(
+              'netC received unexpected data from netA' +
+                '\n  expected: ' +
+                payload.toString('utf-8') +
+                '\n    actual: ' +
+                packet.payload.toString('utf8')
+            );
+          }
+        }
+      });
+
+      netA.connect({
+        name: 'netA-' + uniqueId(),
+      });
+
+      netB.connect({
+        name: 'netB-' + uniqueId(),
+      });
+
+      netC.connect({
+        name: 'netC-' + uniqueId(),
+      });
+
+      return cleanUp;
+    },
+    { timeout: 5000 }
+  );
+});
+
+test.run();
